myproject/requirements.txt
absl-py==2.4.0
asgiref==3.11.1
astunparse==1.6.3
certifi==2026.2.25
charset-normalizer==3.4.7
Django==6.0.3
flatbuffers==25.12.19
gast==0.7.0
google-pasta==0.2.0
grpcio==1.80.0
gunicorn==25.3.0
h5py==3.14.0
idna==3.11
keras==3.14.0
libclang==18.1.1
markdown-it-py==4.0.0
mdurl==0.1.2
ml_dtypes==0.5.4
mysqlclient==2.2.8
namex==0.1.0
numpy==2.4.4
opt_einsum==3.4.0
optree==0.19.0
packaging==26.0
pillow==12.2.0
protobuf==7.34.1
Pygments==2.20.0
requests==2.33.1
rich==14.3.3
setuptools==82.0.1
six==1.17.0
sqlparse==0.5.5
tensorflow==2.21.0
termcolor==3.3.0
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
wheel==0.46.3
wrapt==2.1.2
whitenoise
